Updated class to use pressure as a float
Updated pressure to be a float and included formula into class that converts voltage difference into pressure in PSI

+++ Main_Code/Pressure.h
@@ -3,7 +3,7 @@ class Pressure {
 
   //Class Variables
   char _trigPin;
-  int _Product;
+  float _Product;
 
   public:
 
@@ -13,7 +13,7 @@ class Pressure {
   }
 
 //Reads the Pressure Transducer value and outputs the pressure 
-  int Pressure_Run() {
+  float Pressure_Run() {
     _Product = analogRead(_trigPin);
     /*
      *
